requirements.txt
pymongo>=4.6.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0
pillow>=10.1.0
numpy>=1.24.0
dnspython>=2.4.2
tqdm>=4.66.0
